+++ packages/bridge-controller/src/index.ts
@@ -52,6 +52,8 @@ export type {
   FeatureFlagsPlatformConfig,
 } from './types';
 
+export { AbortReason } from './utils/metrics/constants';
+
 export { StatusTypes } from './types';
 
 export {
@@ -101,7 +103,12 @@ export {
   SWAPS_CHAINID_DEFAULT_TOKEN_MAP,
 } from './constants/tokens';
 
-export { SWAPS_API_V2_BASE_URL } from './constants/swaps';
+export {
+  SWAPS_API_V2_BASE_URL,
+  SWAPS_CONTRACT_ADDRESSES,
+  SWAPS_WRAPPED_TOKENS_ADDRESSES,
+  ALLOWED_CONTRACT_ADDRESSES,
+} from './constants/swaps';
 
 export { MetricsActionType, MetricsSwapType } from './utils/metrics/constants';
 
@@ -156,3 +163,10 @@ export { DEFAULT_FEATURE_FLAG_CONFIG } from './constants/bridge';
 export { getBridgeFeatureFlags } from './utils/feature-flags';
 
 export { BRIDGE_DEFAULT_SLIPPAGE } from './utils/slippage';
+
+export {
+  isValidSwapsContractAddress,
+  getSwapsContractAddress,
+  fetchTokens,
+  type SwapsToken,
+} from './utils/swaps';