+++ packages/subscription-controller/src/types.ts
@@ -171,7 +171,10 @@ export type ProductPrice = {
   /** only usd for now */
   currency: Currency;
   trialPeriodDays: number;
+  /** min billing cycles for approval */
   minBillingCycles: number;
+  /** min billing cycles for account balance check */
+  minBillingCyclesForBalance: number;
 };
 
 export type ProductPricing = {
@@ -273,7 +276,6 @@ export type Cohort = {
 export type SubscriptionEligibility = {
   product: ProductType;
   canSubscribe: boolean;
-  minBalanceUSD: number;
   canViewEntryModal: boolean;
   modalType?: ModalType;
   cohorts: Cohort[];